+++ js/utils.js
@@ -1,0 +1,18 @@
+const getRandomInt = (min, max) =>
+  Math.round(Math.random() * (max - min) + min);
+
+const createRandomIdGenerator = (min, max) => {
+  const previousValues = [];
+
+  return function () {
+    let currentId = getRandomInt(min, max);
+
+    while (previousValues.includes(currentId)) {
+      currentId = getRandomInt(min, max);
+    }
+    previousValues.push(currentId);
+    return currentId;
+  };
+};
+
+export { getRandomInt, createRandomIdGenerator };
